saturation: add CIE Yuv and RGB interpolation methods, issue #171
All three methods are linear interpolations towards luminance desaturation of
the image, though in different color spaces - yielding slightly different
results. The default behavior remains the CIE Lab based interpolation for now.
Perhaps in a later release we will also change the default to be the
interpolation in linear native space, while adding CMYK support to desaturate
that does not lose the separation.