motion-blur-zoom: limit value range to -10.0 to 10.0
There are existing tutorials that expect the user to enter absolute pixel
values rather than relative values which this operation expects. Limiting
to -10.0 to 10.0 encompasses all coordinats within image as well as enough
outside on each side to make it unlikely intentional parameters. The ui
range is already limited to 0.0-1.0

Fixes issue #174