maze: implement depth_first{_tileable} algorithm without recursion
Fix the second issue reported in https://gitlab.gnome.org/GNOME/gegl/issues/77.

Random directions are now obtained by a call to g_rand_int_range instead of the
original hand-written random values computation. So for a given a seed, the
result is not the same as the recursive implementation but very similar.