buffer: in gegl-tile-alloc, limit number of tiles per block
In gegl-tile-alloc, limit the number of tiles in a single block
(currently, to 1024 tiles).  This sets an upper limit for block
overhead even when using a big tile cache, and, in particular,
avoids regularly falling back to single-tile allocations when the
cache size is much higher than memory size (see previous commit).