operation: fix operation-class keys hash aliasing in subclasses
Since we no longer initialize the the keys hash table for each
GeglOperation subclass in base_init(), it is inherited from the
parent class; if the parent class already set some class keys, the
inherited hash table is not NULL, and aliases that of the parent.

Hackishly fix this by storing a pointer to the owner class in the
hash table, and create a new hash table, rather than using the
existing one, in case the stored pointer doesn't match the current
class.