buffer: in gegl_buffer_copy(), avoid copying to/from abyss when using COW
In gegl_buffer_copy(), when copying whole tiles, avoid copying
tiles that coincide with the source or destination abyss, so that
we get the correct abyss behavior as with non-COW copy.  In
particular, this avoids stale cached data in nodes whose bounding
box has changed from leaving traces during processing, when the
cache is being copied from.