Meson port - integrating merge request !35
This commit contains the meson build system changes from is a combination
of commits from that branch, with most of the meson port done by
Félix Piédallu - with some adaptations to sources and build by Øyvind Kolås.

We need a transition phase where meson works on top of autotools to
switch many of the CI systems gegl is involved in over to meson.

Among the commits squashed into this one are:

Author: Félix Piédallu <felix@piedallu.me>
    Remove check for py2
    post-merge
    build: add gegl-npd-lib to pkgconfig
    Added meson files
    [src] move libnsgif to subdir
    [src] Compatibility with Meson: libs/ -> subprojects/

Author: Øyvind Kolås <pippin@gimp.org>
    tools: exp_combine fix signedness warning
    gegl: fix Babl gir name
    meson: do not builds docs and workshop by default
    bin/meson: link binary with gio
    operations/meson: add missing operations
    meson: skip opencl tests, opt all other tests out of parallel running
    put all genrated webcontent in one folder
    docs/meson: build of ops examples/html
    tools: skip searching for source path for ops, adapt for meson
    docs: make asciidoc part of build work with meson, using bits from nielsdg's meson branch
    bin:meson add lua deps
    meson: check for luajit instead of lua
    meson: generate gtk-doc
    meson: drop graphviz checking
    gegl: add back GEGL_MAJOR_VERSION etc in gegl-version.h
    build: add check for malloc_trim