buffer: use binary tree for gap list in the swap backend
In GeglTileBackendSwap, use a binary tree (sorted by gap offset)
combined with a singly-linked list for storing the swap-file gap
list.  While we still use a linear search for finding a large-
enough gap when allocating space for a new tile, this allows us to
use an O (log n) search, instead of a linear search, for the two
adjacent gaps when removing a tile, significantly improving tile-
removal speed.