docs: Depend on operations
Documentation generation relies on programs built in 'tools' directory,
which load modules at runtime from 'operations' directory. While the
dependency on 'tools' is correctly handled by meson because the use of
meson variables, 'operations' is not handled at all. This makes it
possible for 'gallery-stamp' target, which uses 'gegl-tester', to run
before modules in 'operations' are built. 'gegl-tester' fails badly in
this case. It ends up being killed by the kernel because it uses up the
memory on the system.

Fixes: https://gitlab.gnome.org/GNOME/gegl/issues/204