buffer: cast width*height multiplications in _set() and _get() to gsize
This permits larger than INT_MAX sized linear buffers to be set and get
on 64bit systems. On 32bit systems the allocations are likely to fail,
ops relying on reading and writing the whole extent of the source buffer
should be changed to process the data bit by bit. This improves issue #221
and issue gimp#4268 on 64bit systems.