process: destroy node cache, instead of clearing, when new bounding box is smaller
When a node's bounding box changes, if the old tile-aligned
bounding box is not contained within the new tile-aligned bounding
box, destroy the node's cache, instead of clearing the now-out-of-
bounds region.  The problem with the latter is that the node's
bounding box can be arbitrarily large, even without actually having
been rendered; if the bounding box then becomes much smaller,
clearing the difference can take a long time, even if the cache
doesn't have any actual tiles in this region.  Destroying the
cache, on the other hand, is much faster.