buffer: in swap backend, only compress current tile when stalling
In the swap backend, when attempting to push a tile to a full
queue, only compress the current tile before waiting for free queue
space, instead of searching the rest of the queue for additional
tiles to compress.  When the queue is very long, the recurring
linear search itself can end up taking a significant amount of
time, during which the queue must be locked, blocking both other
threads attempting to push a tile simultaneously, as well as the
swap-writer thread attempting to free queue space.

This partially reverts commit
16d65b5d676229a87cc2662a770c73080d018d58.