buffer: in gegl-tile-alloc, keep a spare empty block
In gegl-tile-alloc, when all tiles in a block are freed, keep a
single empty block around at any given time as a spare, and reuse
the block when a new one is required.  This avoids frequent block
allocation and freeing when a single (or few) tiles on the edge of
a block are repeadtedly allocated/freed.