buffer: allow creating buffers with uninitialized data
Add a new construct-only GeglBuffer::initlalized boolean property,
which, when FALSE, allows the buffer to avoid zero-initializing
empty tiles.  This property defaults to TRUE, and is mostly meant
to be used internally for now.

When set to TRUE, empty tiles are created fully damaged, so that
their data is not unnecessarily initialized when uncloned.  Note
that we already avoid initializing or copying uncloned tile data
when overwriting the full tile content, but this allows us to avoid
initialization even when writing partial content.

This optimization is currently limited to level-0 tiles, to
simplify the code.