gegl, operations: add meta-data API
This adds GeglResolutionUnit, GeglMetaDataStore interface and a
hashtable implementation, as suggested and refined in issue #222.

Image loading and saving operations can now implement the "metadata"
property. For loading ops this means extracting known metadata keys when
processing, and for saving ops storing the applicable and set key.  The
file operations can register maps of names of keys between its own
naming scheme and the GeglMetaDataStore naming scheme, as well as
provide conversion functions for rewriting formats, for instance for
timestamps.