buffer: tune cubic coefficients away from anomaly
As reported in issue #167, the current coefficients of the cubic
resampler causes a sinuosidal grid anomaly when rotating some noise
patterns, to work around this we bring the sharpness/smoothness
trade-off for the cubic half way back to smooth which also
significantly reduces the strength of the anomaly.

The code for choosing behavior of the cubic resampler by choosing
cubic types via string never worked, as the strcmp were the inverse
of what they should have been. This commit reduces it to only rely
on computing the coefficent from the formula, and adds comments
what the different constants we have used mean.