variable-blur: new op in common-cxx
Add a new gegl:varaible-blur op, which takes an input and a mask,
and uses the mask as a per-pixel blur factor for the input.  This
is currently implemened by applying a series of Gaussian blurs to
the input with increasing radii, and using gegl:piecewise-blend to
interpolate them using the mask.

The is a C op, but it lives in common-cxx so that it only gets
built if the rest of the C++ ops are built, since
gegl:piecewise-blend is one.