focus-blur: new op in common-cxx
Add a new gegl:focus-blur, which blurs the image around a focal
point, allowing the geometry of the focus region to be adjusted.
It can be used to create fake depth-of-field effects.

This op is simply a front-end for gegl:variable-blur with
gegl:vignette as the mask.  It's a C op that lives in common-cxx
for the same reason as gegl:variable-blur (see last commit).