{variable,focus}-blur: remove levels/gamma props; add high-quality prop
In gegl:varaible-blur and gegl:focus-blur, remove the "levels" and
"gamma" properties, which are implementation details.  Use a fixed
1.5 gamma, so that more levels are allocated to the higher-
frequency scales, and adjust the number of levels logarithmically
according to the blur radius, so that bigger radii use more levels.

Add a "high-quality" boolean property, which can be used to produce
better and more-consistent output across different radii, at the
cost of speed.  This is especially useful to avoid discontinuities
when animating the radius.  This property is off by default.

Allow #defining MANUAL_CONTROL to get the old "levels" and "gamma"
properties back, for testing.