crop: change default values and add inferred defaults
gegl:crop alrready supports specifying the rectangle to crop to by
specifying connecting a node to the aux pad. Often the desired node to
connect ends up being the node at the end of the producer chain along
input pads this is now used  when the default values of 0,0 0x0.

This means that it is now sufficient to do:

gegl-unsharp-mask