transform: round numerical errors in transformation matrix
In the transform ops, round numerical errors in the generated
matrix to the nearest integer.  This is especially important when
using the NEAREST sampler, which is sensitive to boundary
conditions.

In particular, a perfect 90-deg rotation matrix cannot be generated
naively using the trigonometric functions (since there is no double
x such that cos(x) = 1 exactly), which can cause a single row or
column to be missing from the output in some cases.