convert-space: handle CMYK and fix bug in property name
We cannot have a property named "name" since this is a property that
exists on GeglNode. In addition a logic bug made the name always
take precedence, forcing "sRGB" to be the converted to space.

We now also create "YA float", "RGBA float" or "CMYKA float" depending
on the type of color model associated with the space instead of
forcing RGBA, thus gegl:convert-space is sufficient to go between
RGB and CMYK.