operations, meson: make gegl:introspect more robust.
First of all, since "gegl:introspect" is apparently not meant to be an
optional operation, and since it relies on the `dot` program (from
graphviz), we need to search for `dot` in the configuration step (I
could check it exists also on Windows, msys2 even has a package, and it
looks like there are macOS builds and probably BSD too…). If we realize
it is a blocker on some platform, then we should make the whole
operation optional (and unbuilt) rather than leaving it broken.

Also instead of calling `dot` directly, call the program detected at
configuration by passing it through a macro (with an exception for
Windows, where I use "dot.exe" as we will assume everything is on the
same bundled bin/ directory there and build-time paths will be wrong at
runtime).

Thirdly build a temporary filename making sure the file didn't exist
before. On some early tests without the `dot` binary, I even thought it
worked without, but it was only because a file from previous run did
exist in the temporary directory! So let's make proper temp files with
unique names and ensure we delete them after use.

When calling the dot command, do not only check the return value -1
(typically here, I got 127 for a non-existing tool in PATH for
instance), let's consider all non-zero return values as a failure.

Finally when we do consider that `dot` call failed, do not run the GEGL
graph and check that o->user_data is non-NULL before using it. We
therefore avoid various causes of possible crashes.

See also: https://gitlab.gnome.org/GNOME/gimp/-/issues/6045