operations: gegl:introspect to use the new is_available() class method.
"gegl:introspect" depends on the external "dot" tool availability. If
absent, we should hide the operation.

We still search for `dot` on the main configure step, but now only for
informative purpose (and outputting the result on the configuration
summary). It does not block anymore the build and the operation is
always built, whether or not dot is found at configure time.