vector-fill: add color space management
If an input buffer is provided, we inherit its color model and color
space which will be used as device space. The color space of the
specified fill color is passed to ctx as the user space. If no input
buffer is provided the color space of the color parameter is used
as the color space of the output.