distance-transform: Add edge handling / abyss_policy property (#280)
Adds an abyss_policy parameter to the distance-transform op (displayed
as "Edge handling" to the user), allowing users to choose whether areas
outside the input are to be treated as above threshold or below
threshold (i.e. infinitely white or black respectively) for calculating
distance.  This allows the op to calculate distance only from content in
the input, rather than always placing a "border" around the output.
The default behavior is set to "Above threshold", a change from previous
behavior, which always treated the abyss as below threshold.

In the special case of an input that is entirely above threshold, with
the abyss also set to above threshold (essentially producing an infinite
sheet), the entire output distance is set to the input's width + height,
representing an infinite distance.  Any code that may want to account
for this case can test any pixel for a distance equal to or greater than
this value, in case it is changed to something more mathematically
correct or otherwise convenient in the future.