Include opencl_dep as a dependency in gegl_lib target
Follow-up to 392e60be7371362c1149a9825d7644df86416692.

In the above commit, the opencl_dep dependency was declared but
this dependency was never added to the gegl_lib target, which
needs it. As such, the addition was a no-op.

This commit adds the opencl_dep dependency to the gegl_lib target
and keeps only the opencl_headers sources, since the gegl_sources
are already included in the gegl_lib target.