gegl: check that the fallback of operation handlers exist.
The loader fallback in particular is "gegl:magick-load" which depends on
a runtime tool (`convert` from Image Magick) so gegl_has_operation()
would verify that the operation is actually available at runtime (since
previous commit). If unavailable, let's just raise a warning and not
call the fallback, which is better than calling it and raising weird
unavoidable errors.