Replace deprecated AVStream::codec usages with a separate AVCodecContext
Two AVCodecContext objects are created, one for audio stream and one for
video stream. When the audio_st/video_st are created, parameters are
stored in the AVCodecParameters (i.e., codecpar fields) and then they
are moved to the AVCodecContext instances in open_audio/open_video
functions. Some fields must be set directly on the AVCodecContext and
have no counterpart in the AVCodecParameters. Initialization of those
fields is postponed to the open_* functions too.