Replace {first,cur}_dts + Remove no-op functions & unused vars
The cur_dts and first_dts fields are moved to internal.h and the
avformat library is supposed to use them exclusively. The cur_dts
is the last seen packet dts, so it can be replaced by pkt.dts and
the first_dts is the first seen packet dts, so it can be kept in
a new p->first_dts variable after the first decoding (in video stream).

The av_register_all and avcodec_register_all functions have no effect
in the new version.

The coded_frame is removed and it was used for reporting more
information to the caller (which is not required in gegl).
For details, check https://github.com/FFmpeg/FFmpeg/commit/11bc79089378a5ec00547d0f85bc152afdf30dfa

Finally, the filename is replaced by url which is allocated dynamically
with no size limit.