meson: GEGL's npd lib is not private and is actually used by software.
Commit be8978bab demoted libgegl-npd to a private library, not caring
about the fact that this library is actually needed externally. Its API
is meant to be exposed publically, headers are installed and so on. In
particular, it is used by GIMP so this commit actually broke GIMP
builds, as well as our CI.

I can understand that adding this side library this way is not so
conceptually ideal, so one should be welcome to provide alternatives,
such as maybe its own pkg-config file, and also changing the API
namespace to gegl_npd_*() or something like this. But until this is done
and GIMP code and build at least is ported to such updated API, we
should not hide away this library meant to be public.