meson: search for babl-0.1 first AND last.
This is one of these weird workaround stuffs. Basically meson looks for
babl as a subproject only if it didn't find it in the system (this part
makes sense), but it will only look to fill in the required dependency,
which is our latest test, and we did it with the name 'babl', which is
not the current name in HEAD of `master` branch of babl (as declared by
meson.override_dependency()). So meson doesn't find babl as subproject
either and the build fails.

This is why I do the initial search of 'babl-0.1' last, which is
redundant to the first test but only serves to be used as the name which
will be searched as a subproject.

Note that we could have avoided all this with recent meson 0.60.0 which
allows to give several names to dependency(), but we don't want to do
the same error as was done some time ago when we used too recent a meson
requirement for babl.

This should fix the CI.