buffer: fix invalid free in tile_unef, masked by g_slice
Instead of allocating and freeing separate small arrays for
the tile clone bookeeping use integers in the GeglTile struct
that do not need to be freed - thus not needing to keep track
of whether the n_clones pointer is an allocation or not.

Before this commit we g_slice would silently mask our attempts at freeing
invalid pointers to small segments taken out of larger slices.

Fixing issue #320