buffer: maintain clone count with tile_data
Partial reversion of preceding commit, the previous commit could lead to
problems when the first tile among clones is not the last. By using the
previous logic for allocating the n_clones array along with the tile data
the life times of the n_clones array matches what it is tracking.