gegl: new API gegl_node_connect
The mental overhead of thinking about sinks and sources and which
API to use of gegl_node_connect_from and gegl_node_connect_to is
solved by introducing a third option, which sniffts out which pad
is input and output pad and makes the connection accordingly.

Deprecation warning added for gegl_node_connect_from which suggests
using gegl_node_connect.