bevel: add distance-transform mode, improve ui, rename to chamfer
Add a more direct analytical bevel that scales with the dimensions
of used text and has more directly accesible controls. The visual
fidelity of the non-distance-transform mode is often smoother
or can has a more interesting shape. Being able to control the
blurs based approach with the same/similar controls would be nice.

There are currently third-party operations in circulation
polluting the gegl: namespace, avoid colliding for now
giving time for other users of gegl:bevel to update,
gegl:bevel as a name should be available for GEGL itself
to use though, and migrating to gegl:bevel for this
operation or a derivative in the future is the plan.

See issue #329