gegl: new gegl_color_get_rgba_with_space() and gegl_color_get_rgba_with_space().
The current get|set_rgba() functions returns or set values as linear
light non premultipled RGBA data (within the sRGB space), which is very
limited. We may want to get or set values within another space, or maybe
simply even within the space this GeglColor was set as. These new
functions allow it.

As a few other functions, it is possible to use a Babl format instead of
the space argument, and the function will take care of using the space
only to return values in "R'G'B'A float" within this space.