hue-chroma: increase epsilon
As discovered in gimp#10865 we are inventing colors for achromatic pixels,
the op already contained an epsilon, increasing it by two orders of magnitude
to 0.0001 seems to match the precision of which the RGB to Lch conversion
maintains achromatic pixels. This is still an order of magnitude more precise
than 1.0/65536.