gegl: better error reporting when parsing gegl chains
Replace all calls to gegl_node_connect with a local to the file
gegl_node_connect_safe, which sets a error message for the UI if pads do not
exist. This should stop the crash reporting in GIMP for these calls; when other
code in GIMP or gegl tries connecting we still want the more expensive crash
reports. This should stop the worst of #360.