Issue #382: fix crash because of non-initialized node used in a graph.
Reviewer's (Jehan) notes: as explained in #382, swapreplaceontop node was
uninitialized but with this patch, it is instead always NULL, which
means it is useless anyway. Also it means that in the first
gegl_node_link_many(), the graph is always cut short right at this node
since NULL is the end flag, which implies the right-part in the node
chain is currently bogus (it is never connected).

Now LinuxBeaver says the op works fine. Maybe the needed parts of the
graph are still reconnected later on (there are a bunch of
gegl_node_link_many() and gegl_node_connect()), maybe these were just
unneeded to start with… What is sure is that the code in its current
state has what looks like moot code here and there which could be
cleaned out.

I still leave the patch as-is because I can't make the time to review
the whole op's logic, and at least it does definitely fix a crash with
no logic change. So let's go with it. Yet this is clearly not optimal.