operations: fix typo s/"sinusodial"/"sinusoidal"/.
Right now, the nick of an enum cannot really be considered part of the
API, AFAIK. Yet from GIMP 3.0 and forward, we generate choice strings
from these for arguments of the libgimp API to create layer effects.

In other words, you could set the "type" argument to "sinusodial"
string. So it's better to fix such typos before GIMP 3.0 goes stable!
From this point in time, we'll consider these nicks part of the stable
API (at least as far as GIMP is concerned).