opencl: fix #406 invalid build options on ARM device
Before using option -cl-fp32-correctly-rounded-divide-sqrt
the docs say we need to test for flag
CL_FP_CORRECTLY_ROUNDED_DIVIDE_SQRT being present, which we did not
do.

This commit adds that check and while at it, also make sure that
the OpenCL C compiler also supports 3.0, which is supposedly not
always the case.

While at it, also reduce the size of the char arrays for name and
version. They are never that long and we set a size, so it won't
overrun.