gegl: mark GeglNode as inert in dispose
This avoids the node being taken into account for get_producer(),
more locations could be added if other similar use-after free
could be triggered by invalidation events or similar. For now this
fixes gegl chains on the commandline of the type:

 $ gegl in.png -o out.png -- dropshadow crop

The graph would be correctly processed but segfault during
cleanup-phase. This issue manifested under alpine/postmarketos but not
under debian.