subprojects: Drop unmaintained Meson subprojects
Partially reverts 5e75afe0

This auto-downloading approach is problematic even for a fallback,
in many ways:

1) such auto-downloaded subprojects were forgotten to be updated;

2) as a result of lack of of maintenance, some fails to build.
babl.wrap, by the way, always fail on Windows but babl repo not;

3) this approach it is not clear enough for most packagers. So,
for example, if due to a timeout on MSYS2 servers json-glib is
not installed, an old version will be downloaded and built without
the packager's consent, which only will be aware by checking the
logs or, more likely (as demonstrated above), if the build fails.
This is the main reason for dropping. I always was annoyed by
that because this affected GIMP CI sometimes and not even the
maintainer was understanding what happened (since, as said, this
is not clear enough), but OpenEXR, after libdeflate and openjph,
made clear how this is more prone to broke than helpful at all:
https://github.com/AcademySoftwareFoundation/openexr/issues/2123

Also, even if we maintain that, why? Considering the rationale of
the MR, that need does not holds up anymore, since such deps are
available on MSYS2 and VCPKG. So, let's not try to do packagers job.