operations: Redraw NDE Kaleidoscope across image
If you use Kaleidoscope as a NDE filter and draw on a
sufficient large image, the filter won't update the whole
image when you draw unless you force a refresh of the
filter. This results in the Kaleidoscope not mirroring the
brush strokes correctly.

El Tomito encountered a similar issue in their custom
GEGL filter, and resolved it with commit
https://gitlab.com/eltomito/eschertile/-/commit/da9fcdd293990beae2eb52137ec2bc24a384aedc
This patch copies that implementation for mirrors.c.
It checks if the change to the filter's input is within bounds,
and returns either just that area or the whole area for rerendering.